requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
click==7.1.2
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1
django-background-tasks==1.2.5
django-common-helpers==0.9.2
django-compat==1.0.15
django-cron==0.5.1
djangorestframework==3.11.1
gunicorn==20.0.4
pathspec==0.8.0
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
psycopg2==2.8.5